ALL - Implemented core temperature fail-safe and preview - Added ability to save results of a hash session and its metadata - Added Wall and Pedantic compiler tags, refactored to appease them - Refactored to include Session struct for user-provided metadata bundling - Changed template assertion to C++20 required concept - Bug fixes
UNIX
- Added /sys/class/thermal file walker to discover all tracked temperatures for use in fail-safe

WINDOWS
- Nothing as of this update

+++ src/foundation/numberSystem.h
@@ -23,7 +23,7 @@ public:
     inline constexpr static void setBase(int b) noexcept { base = b; }
     inline static auto getBase() noexcept -> int { return base; }
 private:
-    inline static int base{};
+    inline static ull base{};
 
 };
